Roll memoization directly into TypedScopeCreator.
This is a prerequisite for handling block scoping correctly: to ensure names are defined in the correct order, TypedScopeCreator needs to create nested block scopes when they are encountered in the createScope traversal, rather than afterwards.  This is required for, e.g. `if (true) { class Base {} } class Sub extends Base {}`.  TypedScopeCreator can now create these scopes and be sure they are correctly memoized.  This is straightforward since every usage of TypedScopeCreator was always wrapped in MemoizedTypedScopeCreator.  The upshot is that MemoizedTypedScopeCreator can now be deleted.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187245381