Extract ScopedName as an abstract superclass of AbstractVar.
Background: CL 181680335 fixed/generalized the equals and hashCode behavior of Var and TypedVar, which previously only depended on the definition node, to instead depend on the name and the root node of the scope in which the name was defined.  Importantly, this preserved the property that Vars needn't share the same scope creator to be equal, but was still required to be an instance of AbstractVar, which requires a corresponding AbstractScope.  This CL loosens that requirement, in particular by providing a simple value type `ScopedName.of(name, node)` which can be used alongside (e.g.) TypedVar as keys in the same ScopedName-keyed map.  This is a step toward supporting block scopes in OTI (#2796).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187266949