Refactor LinkedFlowScope to key off of ScopedName instead of String.
This is a step toward teaching TypeInference to deal with ES6 block scopes correctly (#2796).  It doesn't make any functional difference yet since all the vars are currently in the same (CFG root) scope, so the comparisons might as well be strings.  The next step is to add a block scope field to LinkedFlowScope in addition to the functionScope field, which will be used by the infer and getSlot methods to map strings to variables in the relevant scope.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187412071