Tighten the strict missing property check.
This tightens the strict missing property check:
A warning is now emitted when assigning a property on a non-`struct` type when that property is not defined on that type.

The loose missing property checks would only emit warnings if the type of the resulting expression was "unknown" now they are emitted regardless of the type of the value.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187514949